requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.2
django-admin-honeypot==1.1.0
django-autoslug==1.9.9
django-recaptcha==3.0.0
django-settings-export==1.2.1
filelock==3.12.0
Flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql==0.0.3
mysql-connector-python==8.0.32
mysqlclient==2.1.1
platformdirs==3.5.0
protobuf==3.20.3
PyMySQL==1.0.3
requests==2.30.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.0
Werkzeug==2.2.3
